# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban API",
    "version": "v1",
    "summary": "Read and change a board from a script.",
    "description": "Hand-written, not generated. Two tests hold it to the code: one fails when a route is registered and missing here or the other way round, one validates every handler response in the test suite against the schema below.\n\nThere are no tokens. The API grants what the board grants, which is everything to whoever can reach the port, so a deployment puts the same proxy or Cloudflare Access in front of it. Where identity is configured it is used for two things: \"@me\" as an assignee, and the author of a comment.\n\nA POST or PUT body must be application/json. Unknown fields are refused with a 400 that names the field, because \"titel\" for \"title\" would otherwise store a card with no title and nobody would find out until they read the board."
  },
  "servers": [
    {
      "url": "/",
      "description": "The board itself. The paths below are absolute, so they work against any host that serves it."
    }
  ],
...
  "paths": {
...
    "/api/v1/boards": {
      "get": {
        "operationId": "listBoards",
        "tags": [
          "boards"
        ],
        "summary": "Every board.",
        "responses": {
          "200": {
            "description": "The boards, each with its columns and labels.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createBoard",
        "tags": [
          "boards"
        ],
        "summary": "Create a board.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The board, with the columns it was created with.",
            "headers": {
              "Location": {
                "description": "Where the new board can be read.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "A board with that slug exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "get": {
        "operationId": "getBoard",
        "tags": [
          "boards"
        ],
        "summary": "One board, with its columns and labels.",
        "description": "The cards are a separate call, because a board with a thousand cards on it is not a response anybody wanted from a request for the column names.",
        "responses": {
          "200": {
            "description": "The board.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "renameBoard",
        "tags": [
          "boards"
        ],
        "summary": "Change the name, the slug, or both.",
        "description": "A field left out keeps its current value. Changing the slug changes the URL of the board and of its page, and the old one stops working.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The board as it is now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Another board already has that slug.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteBoard",
        "tags": [
          "boards"
        ],
        "summary": "Delete a board and everything on it.",
        "description": "Columns, labels, cards, comments. There is no undo and no archive for a board.",
        "responses": {
          "204": {
            "description": "Gone."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/layout": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "put": {
        "operationId": "setLayout",
        "tags": [
          "boards"
        ],
        "summary": "Columns side by side, or stacked as rows.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Set."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "board": {
        "name": "board",
        "in": "path",
        "required": true,
        "description": "The board's slug, the same one its page URL uses.",
        "schema": {
          "type": "string"
        },
        "example": "board"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed, a field is unknown, or the service refused a value. field names the offending input where there is one.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No such board, column, label, card or comment. A column or label id that exists on another board is also this.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Content-Type is set to something other than application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error": {
        "description": "Every failure has this body, including a 404 for an endpoint that does not exist and a 500.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in one sentence."
          },
          "field": {
            "type": "string",
            "description": "The input the service refused, when it named one. Absent otherwise."
          }
        }
      },
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "name",
          "layout",
          "sla",
          "columns",
          "labels",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "What the URLs use."
          },
          "name": {
            "type": "string"
          },
          "layout": {
            "type": "string",
            "enum": [
              "columns",
              "rows"
            ],
            "description": "How the page draws it. Read as columns where it was never set."
          },
          "sla": {
            "$ref": "#/components/schemas/SLA"
          },
          "columns": {
            "type": "array",
            "description": "In board order, left to right.",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Label"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Column": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "position",
          "wip_limit",
          "stops_clock"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "description": "1 is the first column."
          },
          "wip_limit": {
            "type": "integer",
            "description": "0 for a column that holds as many cards as you like."
          },
          "stops_clock": {
            "type": "boolean",
            "description": "A column where the response-time clock does not run, such as Done."
          }
        }
      },
      "SLA": {
        "type": "object",
        "additionalProperties": false,
        "description": "How long a card may sit untouched during office hours.",
        "required": [
          "response_hours",
          "days",
          "start",
          "end",
          "zone"
        ],
        "properties": {
          "response_hours": {
            "type": "integer",
            "description": "Office hours, not hours on the wall. 0 for a board with no promise."
          },
          "days": {
            "type": "array",
            "description": "The office days, Monday first.",
            "items": {
              "type": "string",
              "enum": [
                "mon",
                "tue",
                "wed",
                "thu",
                "fri",
                "sat",
                "sun"
              ]
            }
          },
          "start": {
            "type": "string",
            "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]$",
            "description": "When the desk opens, in zone."
          },
          "end": {
            "type": "string",
            "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]$",
            "description": "When it closes. 00:00 is the midnight that ends the day."
          },
          "zone": {
            "type": "string",
            "description": "An IANA name such as Europe/Zurich. Empty is UTC."
          }
        }
      },
      "Label": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "color"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "description": "A hex colour, lower case. Empty for the default the page picks."
          }
        }
      },
...
      "BoardInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "Left out, it is derived from the name."
          },
          "columns": {
            "type": "array",
            "description": "The columns to create with the board. Left out, the board gets To Do, In Progress and Done.",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        }
      },
      "BoardPatch": {
        "type": "object",
        "additionalProperties": false,
        "description": "A field left out, or sent as null, keeps its current value. Sending \"\" is a value and the service refuses it.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 100
          },
          "slug": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          }
        }
      },
      "LayoutInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "layout"
        ],
        "properties": {
          "layout": {
            "type": "string",
            "enum": [
              "columns",
              "rows"
            ]
          }
        }
      },
...
    }
  }
}